Change decision criterion for correlation, update notebooks
Instead of marking a relation with a low p-value AND "high enough"
r-value as a correlation, only p-value is now used as a criterion and
the r-value is left for interpretation to the user.

Notebooks were updated accordingly, including some other minor changes
in the commentary. The outputs are now included as well.